KernelArray RAII (#25)
* move ArraySize2D to its own header

* move ArraySize2D operators to array_size_2d.hpp

* change dims data type from int to size_t for cuda; still some int dims in mpi

* create KernelArrayView but not using yet

* kernels now use KernelArrayView (instead of KernelArray)

* implement KernelArray with constructor and View method

* all set with KernelArray RAII except cuda solver needs to use its KernelArray member instead of its int* device_data_ member

* KernelArray RAII fully implemented

* change MultiTileKernel edge buffers to KernelArray in launcher and KernelArrayView in actual kernel

* remove unused code

+++ src/common/arg_parser/arg_parser.cpp
@@ -18,11 +18,11 @@ ProgramArgs ArgParser::Parse(int argc, char *const argv[]) {
   int seed = 1234;
   std::string backend = "mpi";
   std::string log_level = "warning";
-  std::vector<int> full_matrix_dim = {4, 4};
-  std::vector<int> tile_dim = {2, 2};
+  std::vector<size_t> full_matrix_dim = {4, 4};
+  std::vector<size_t> tile_dim = {2, 2};
 
   // Optional CUDA-specific args
-  std::optional<std::vector<int>> sub_tile_dim;
+  std::optional<std::vector<size_t>> sub_tile_dim;
   std::optional<std::string> cuda_kernel;
 
   // Registered CLI options
@@ -92,7 +92,7 @@ ProgramArgs ArgParser::Parse(int argc, char *const argv[]) {
     }
 
     if (!subtile_option->count()) {
-      sub_tile_dim = std::vector<int>{2, 2};  // Default subtile size
+      sub_tile_dim = std::vector<size_t>{2, 2};  // Default subtile size
     }
   }
