+++ test/mpi_prefix_sum/mpi_environment/test_mpi_environment.cpp
@@ -7,9 +7,9 @@
 class MpiEnvironmentTest : public ::testing::Test {
 protected:
 
-  std::vector<int> full_matrix_dim_ = std::vector<int>({6, 6});
-  std::vector<int> tile_dim_ = std::vector<int>({3, 3});
-  std::optional<std::vector<int>> maybe_subtile_dim_ = std::nullopt;
+  std::vector<size_t> full_matrix_dim_ = std::vector<size_t>({6, 6});
+  std::vector<size_t> tile_dim_ = std::vector<size_t>({3, 3});
+  std::optional<std::vector<size_t>> maybe_subtile_dim_ = std::nullopt;
   std::optional<std::string> maybe_kernel_ = std::nullopt;
 
   ProgramArgs program_args_ = ProgramArgs(
